+++ packages/react/src/atomic/MultiSelect/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 Use a multi-select box when a user needs to select more than one option from a list.
 
+## 2026-04-06 `1.2.2`
+
+#### 🐛 Bug fixes
+
+- Trigger border radius matches `Select` (`radius.base`, 6px) via `@nimbus-ds/styles`. ([#455](https://github.com/TiendaNube/nimbus-design-system/pull/455) by [@noecondoleo](https://github.com/noecondoleo))
+
 ## 2026-02-12 `1.2.0`
 
 #### 🎉 New features